Add spdlog logging

+++ src/test_libfirestorm/src/DotProduct.cpp
@@ -7,7 +7,7 @@
 DotProduct::DotProduct()
     : vector_a{32}, vector_b{32}, result(1.5)
 {
-    for (int i = 0; i < vector_a.dimensions; ++i) {
+    for (size_t i = 0; i < vector_a.dimensions; ++i) {
         vector_a.data[i] = 0.5;
     }
